:recycle: #553 - refactor: make settings page a nested route
