+++ frontend/src/pages/settings/index.ts
@@ -1,3 +1,1 @@
-export * from "./ShortProcedureSettingsPage";
-export * from "./ShortProcedureSettingsPage.action";
-export * from "./ShortProcedureSettingsPage.loader";
+export * from "./short-procedure";
