+++ frontend/src/pages/settings/short-procedure/index.ts
@@ -1,0 +1,3 @@
+export * from "./ShortProcedureSettingsPage";
+export * from "./ShortProcedureSettingsPage.action";
+export * from "./ShortProcedureSettingsPage.loader";
